Move callback and actions out of BubbleActions.java
